move sharedProps, styleoverride file insider shared folder
