+++ docs/src/componentDocs/Drawer/markdown/DrawerAPIDocs.mdx
@@ -2,7 +2,7 @@ import Box from '@mui/material/Box';
 
 import Drawer from '../images/drawer.png';
 import DrawerAnatomy from '../images/drawerAnatomy.png';
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer
