+++ docs/src/componentDocs/DrawerBody/markdown/DrawerBodyAPIDocs.mdx
@@ -1,5 +1,5 @@
 import Box from '@mui/material/Box';
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer Body
