+++ docs/src/componentDocs/DrawerLayout/markdown/DrawerLayoutAPIDocs.mdx
@@ -1,6 +1,6 @@
 import Box from '@mui/material/Box';
 import DrawerLayoutImage from '../images/drawerLayout.png';
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer Layout
