+++ docs/src/componentDocs/DrawerNavGroup/markdown/DrawerNavGroupAPIDocs.mdx
@@ -1,6 +1,6 @@
 import Box from '@mui/material/Box';
 import DrawerNavGroupAnatomy from '../images/drawerNavGroupAnatomy.png';
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer Nav Group
