+++ docs/src/componentDocs/DrawerNavItem/markdown/DrawerNavItemAPIDocs.mdx
@@ -1,7 +1,7 @@
 import Box from '@mui/material/Box';
 import NavItemAnatomy from '../images/navItemAnatomy.png';
 
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer Nav Item
