+++ docs/src/componentDocs/DrawerRailItem/markdown/DrawerRailItemAPIDocs.mdx
@@ -1,7 +1,7 @@
 import Box from '@mui/material/Box';
 
 import RailAnatomy from '../images/railAnatomy.png';
-import SharedProps from '../../../markdown/SharedProps.mdx';
+import SharedProps from '../../../shared/markdown/SharedProps.mdx';
 
 # Drawer Rail Item
