+++ docs/src/shared/markdown/SharedProps.mdx
